break resolving reference preprocessing into its own class so it can be reused
